Extract node module
